Add a targets-field to the ExecEngine and to a filter on the command; append the arguments to the command if the target name matches the crate name.
This approach is probably too naive. It should be possible to filter based on e.g. a pkgid instead of a name,
can it ever happen that a dependency have the same name as the package?
What happens if we get a name conflict somewhere, does the command provide some more metadata that can be used to ensure we append to the correct command?